Add `label` to `Association`
This change adds support for a new optional field on the association
mixin, used by `definition`, `footnoteDefinition`, the `Reference`
mixin (in turn used by `linkReference`, `imageReference`, and
`footnoteReference`).

The value of the `label` field must, when present, hold the
non-normalised value of the `identifier` field.

This allows references that are not association with definitions
to be represented correctly as raw text again.

Closes GH-23.
Related to syntax-tree/mdast-util-to-hast#20
Related to syntax-tree/mdast-util-to-hast#21
Related to syntax-tree/mdast-util-to-hast#22
Related to remarkjs/remark#346